client/requirements.txt
PyAudio==0.2.11
python-socketio[client]
sounddevice==0.3.14
